Add makefile, fix Problem006.
Add a makefile. In the future it should use matching, but right now it
is just brute forcing it.

Problem006 is interesting; it had an uninitialized value in the for loop
(short i; ...). i still properly counted from 0 to 100, but the
variables sumOfSqr and sum (which properly incremented inside the loop)
reset to 0 when i went out of scope. Odd.